--- app/src/types/git.ts
@@ -1,1 +1,0 @@
-export type CommitHash = string;
