--- app/src/types/modal.ts
@@ -1,5 +1,0 @@
-export enum ModalType {
-  NewFile = 'newFile',
-  DeleteFile = 'deleteFile',
-  CommitMessage = 'commitMessage',
-}
