--- app/src/types/theme.ts
@@ -1,4 +1,0 @@
-export enum Theme {
-  Light = 'light',
-  Dark = 'dark',
-}
